create View class for twig
