+++ controller/Controller.php
@@ -4,8 +4,6 @@
  * This file contains Controller class
  */
 namespace controller;
-use Twig;
-use Twig_Extensions_Extension_Text;
 
 /**
  * Class for init Twig, send email and control session
@@ -20,7 +18,6 @@ use Twig_Extensions_Extension_Text;
  */
 class Controller
 {
-    protected $twig;
     protected $user;
     protected $msg;
 
@@ -48,19 +45,6 @@ class Controller
         } else {
             $this->user = null;
         }
-    }
-
-    /**
-     * Init twig
-     * 
-     * @return void
-     */
-    public function twigInit()
-    {
-        $loader = new Twig\Loader\FilesystemLoader('view');
-        $this->twig = new Twig\Environment(
-            $loader, ['cache' => false, 'debug' => true]
-        ); 
     }
 
     /**
@@ -115,16 +99,12 @@ class Controller
      * 
      * @return void
      */
-    public function errorView($exception)
+    public function error($exception)
     {
-        $this->twigInit();
-        $this->twig->addExtension(new Twig\Extension\DebugExtension); //think to delete this line
-        $this->twig->addExtension(new Twig_Extensions_Extension_Text()); 
+        $view = new \view\View;
 
-        echo $this->twig->render(
-            'errorView.twig', array(
-                'exception' => $exception 
-            )
-        );
+        $data = array('exception' => $exception);
+        $page = 'errorView.twig';
+        $view -> displayPage($data, $page);
     }
 }
